requirements.txt
python-dotenv==0.19.0
python-telegram-bot==13.7
redis==3.5.3
vk-api==11.9.4